project_tests/gtrobot2-project/gtrfx-backend-main/requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.3.0
asgiref==3.8.0
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
Django==4.1
django-jazzmin==2.6.1
django-modeltranslation==0.18.11
djangorestframework==3.15.1
eskiz-sms==0.2.3
frozenlist==1.4.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.6
MouseInfo==0.1.3
multidict==6.0.5
netaddr==1.2.1
numpy==1.26.4
pillow==10.2.0
PyAutoGUI==0.9.54
pyclick==0.0.2
PyGetWindow==0.0.9
PyJWT==2.8.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-click==0.1.4
python-dotenv==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.31.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
twilio==9.0.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wrapt==1.16.0
xlib==0.21
yarl==1.9.4
yookassa==3.0.1
django-cors-headers==4.3.1
pytz==2024.1